requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
python-dotenv
openai
requests